initvals: fix printf format warnings
The patch fixes the following warnings
(with CFLAGS=-Wformat):

initvals.c: In function ‘ar5008_hw_print_initvals’:
initvals.c:393: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’
initvals.c:394: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’
initvals.c:395: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’
initvals.c:396: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’
initvals.c:397: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’
initvals.c:398: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’
initvals.c:399: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’
initvals.c:400: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’
initvals.c:401: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’
initvals.c:402: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’
initvals.c:403: warning: format ‘%ld’ expects type ‘long int’, but argument 2 has type ‘unsigned int’
...

+++ initvals.c
@@ -257,7 +257,7 @@ typedef long long unsigned int u64;
 						ARRAY_SIZE(_array), ARRAY_SIZE((_array)[0])); \
 		printf("0x%016llx        "#_array"\n", chksum); \
 	} else { \
-		printf("static const u32 "#_array"[][%ld] = {\n", ARRAY_SIZE((_array)[0])); \
+		printf("static const u32 "#_array"[][%d] = {\n", (int) ARRAY_SIZE((_array)[0])); \
 		ath9k_hw_print_initval((const u32 *) _array, \
 				       ARRAY_SIZE(_array), ARRAY_SIZE((_array)[0]), false); \
 	} \
